fraud-inference/requirements.txt
fastapi
uvicorn
joblib
pandas
xgboost
scikit-learn
kafka-python
requests
